+++ example/android/app/src/main/kotlin/de/mintware/flutter_p2p_example/MainActivity.kt
@@ -3,5 +3,5 @@ package de.mintware.flutter_p2p_example
 import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.engine.FlutterEngine
 
-class MainActivity : FlutterActivity() {
+class MainActivity: FlutterActivity() {
 }