+++ generate_proto.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-protoc --dart_out=./lib/gen ./protos/protos.proto
-protoc --swift_out=./ios/Classes ./protos/protos.proto
-flutter format ./lib/gen